% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxalist.R
\name{alltaxalist}
\alias{alltaxalist}
\title{Return all taxa of all ranks in a dataset}
\usage{
alltaxalist(data_m)
}
\arguments{
\item{data_m}{the dataset}
}
\value{
all taxa and ranks of a dataset
}
\description{
Return all taxa of all ranks in a dataset
}
\examples{
\dontrun{
taxalist(data_m)
}
}
