% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_m.R
\docType{data}
\name{data_m}
\alias{data_m}
\title{Example data for biotaxa package}
\format{A data frame with 44915 rows and 8 variables:
\describe{
  \item{AphiaIDs}{AphiaIDs defined in WoRMS}
  \item{Kingdoms}{Kingdoms in WoRMS}
  \item{Phyla}{Phyla in WoRMS}
  \item{Classes}{Classes in WoRMS}
  \item{Orders}{Orders in WoRMS}
  \item{Families}{Families in WoRMS}
  \item{Genera}{Genera in WoRMS}
  \item{year}{Publication year of the taxa group}
}}
\usage{
data_m
}
\description{
A dataset containing the taxonomic details of almost 45,000 species matched by WoRMS in Antarctica and the Southern Ocean.
taxa_m.
}
\keyword{datasets}
