% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBIS_imprecision.R
\name{OBIS_imprecisionrate}
\alias{OBIS_imprecisionrate}
\title{Return the frequency ranking of an interested lower taxon based on a higher taxon}
\usage{
OBIS_imprecisionrate(taxa)
}
\arguments{
\item{taxa}{A string.}
}
\value{
the percentage of taxa occurrences on OBIS
}
\description{
Return the frequency ranking of an interested lower taxon based on a higher taxon
}
\examples{
\dontrun{
OBIS_imprecisionrate("Abra")
}
}
