% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_query.R
\name{rank_query}
\alias{rank_query}
\title{Query and return the rank of a taxa}
\usage{
rank_query(in_taxa)
}
\arguments{
\item{taxa}{A string.}
}
\value{
the basic taxonomic information of the interested taxa.
}
\description{
Query and return the rank of a taxa
}
\examples{
\dontrun{
rank_query("Salpida")
}
}
