% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spetaxalist.R
\name{spetaxalist}
\alias{spetaxalist}
\title{Return all taxa of a specific rank in a dataset}
\usage{
spetaxalist(data_m, rank)
}
\arguments{
\item{data_m}{the dataset}

\item{rank}{the rank; the rank should not be lower than Genus}
}
\value{
all taxa of a rank in a dataset
}
\description{
Return all taxa of a specific rank in a dataset
}
\examples{
\dontrun{
taxalist(data_m, rank)
}
}
