% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa_rich.R
\name{taxa_rich}
\alias{taxa_rich}
\title{Return the predicted value of taxa richness (of given rank) based on logistic regression model}
\usage{
taxa_rich(taxa, rank)
}
\arguments{
\item{taxa}{A string.}

\item{rank}{A string.}
}
\value{
the predicted value of taxa richness (of given rank) based on logistic regression model
}
\description{
Return the predicted value of taxa richness (of given rank) based on logistic regression model
}
\examples{
\dontrun{
taxa_rich("Animalia", "Phylum")
}
}
