% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxaaccum.R
\name{taxaaccum}
\alias{taxaaccum}
\title{Report the accumulative numbers of a rank of a given taxa overtime}
\usage{
taxaaccum(taxa, rank)
}
\arguments{
\item{taxa}{A string.}

\item{rank}{A string.}
}
\value{
accumulation curve of \code{rank} of a \code{taxa}
}
\description{
Report the accumulative numbers of a rank of a given taxa overtime
}
\examples{
\dontrun{
taxaaccum("Animalia", "Phylum")
}
}
