% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxalist.R
\name{taxalist}
\alias{taxalist}
\title{Return all taxa in a dataset}
\usage{
taxalist(data_m)
}
\arguments{
\item{data_m}{the dataset}
}
\value{
all taxa and ranks of a dataset
}
\description{
Return all taxa in a dataset
}
\examples{
\dontrun{
taxalist(data_m)
}
}
