% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxamodel.R
\name{taxamodel}
\alias{taxamodel}
\title{Fitting the taxonomic accumulation of a given taxa with user-selected logistic or Michaelis-Menten model}
\usage{
taxamodel(taxa, rank, method)
}
\arguments{
\item{taxa}{A string.}

\item{rank}{A string.}

\item{method}{A string.}
}
\value{
an accumulative curve of taxa overlapped with a fitting curve of selected model
}
\description{
Fitting the taxonomic accumulation of a given taxa with user-selected logistic or Michaelis-Menten model
}
\examples{
\dontrun{
taxamodel("Animalia", "Phylum", "logistic")
}
}
