% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxamodel_cor.R
\name{taxamodel_cor}
\alias{taxamodel_cor}
\title{Report the correlation coefficient of selected model}
\usage{
taxamodel_cor(taxa, rank, method)
}
\arguments{
\item{taxa}{A string.}

\item{rank}{A string.}

\item{method}{A string.}
}
\value{
the correlation coefficient of the \code{taxa} ~ \code{rank} \code{method} model.
}
\description{
Report the correlation coefficient of selected model
}
\examples{
\dontrun{
taxamodeo_cor("Animalia", "Phylum", "logistic")
}
}
