% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxamodel_FIXPLOT.R
\name{taxamodel_FIXPLOT}
\alias{taxamodel_FIXPLOT}
\title{Report the accumulative numbers of a rank of a given taxa overtime}
\usage{
taxamodel_FIXPLOT(taxa, rank, method)
}
\arguments{
\item{taxa}{A string.}

\item{rank}{A string.}

\item{method}{A string.}
}
\value{
modeling result of the accumulation of \code{rank} of a \code{taxa} by \code{method}
}
\description{
Report the accumulative numbers of a rank of a given taxa overtime
}
\examples{
\dontrun{
taxamodel_FIXPlOT("Animalia", "Genus", "logistic")
}
}
