% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxamodel_test.R
\name{taxamodel_test}
\alias{taxamodel_test}
\title{To cross-test this specific model function}
\usage{
taxamodel_test(taxa, rank, method)
}
\arguments{
\item{taxa}{A string.}

\item{rank}{A string.}

\item{method}{A string.}
}
\value{
the correlation coefficient of a non-linear model of \code{taxa} ~ \code{rank}
}
\description{
To cross-test this specific model function
}
\examples{
\dontrun{
taxaamodel_test("Animalia", "Phylum", "logistic")
}
}
