% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxaprecision.R
\name{taxaprecision}
\alias{taxaprecision}
\title{Report the percentage of unique taxonomic records of a given taxa identified to species level}
\usage{
taxaprecision(taxa)
}
\arguments{
\item{taxa}{A string.}
}
\value{
the percentage of taxonomic records identified to species level
}
\description{
Report the percentage of unique taxonomic records of a given taxa identified to species level
}
\examples{
\dontrun{
taxaprecision("Achnanthes")
}
}
