% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBIS_imprecision.R
\name{OBIS_imprecisionrate}
\alias{OBIS_imprecisionrate}
\title{Return the percentage of taxa occurrences unidentified to the finest taxa level (i.e. species)}
\usage{
OBIS_imprecisionrate(taxa)
}
\arguments{
\item{taxa}{A string.}
}
\value{
the percentage of taxa occurrences on OBIS
}
\description{
Return the percentage of taxa occurrences unidentified to the finest taxa level (i.e. species)
}
\examples{
\dontrun{
OBIS_imprecisionrate("Abra")
}
}
